Eventemitter entfernt da er für die Animation nicht mehr funktioniert
