+++ src/Interfaces/CharacterProps.ts
@@ -1,0 +1,5 @@
+import Richtung from "../Types/Richtung";
+
+export type CharacterProps = {
+    richtung: Richtung;
+  };
