+++ src/Interfaces/IGameFieldProps.ts
@@ -1,0 +1,11 @@
+import { BehaviorSubject } from "rxjs/internal/BehaviorSubject";
+import Richtung from "../Types/Richtung";
+
+export default interface IGameFieldProps {
+    fields: React.FC[][];
+    spielFeldService: {
+      spielFeldSubject: BehaviorSubject<React.FC<{}>[][]>;
+      bewegungsRichtungHackmanSubject: BehaviorSubject<Richtung>;
+      handleKeyDown: (e: React.KeyboardEvent) => void;
+      }
+  }
